requirements.txt
pyserial==3.5
PyQt6==6.7.0
